requirements.txt
streamlit==1.45.0
pandas==2.0.1
numpy==1.24.4
plotly==6.0.1
scikit-learn==1.3.0
google-generativeai==0.5.3
sqlalchemy==2.0.0
psycopg2-binary==2.9.7
pillow==9.4.0
matplotlib==3.7.1
scipy==1.11.2
imbalanced-learn==0.11.0
joblib==1.3.2 